requirements.txt
Flask==2.2.3
matplotlib==3.6.3
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
scikit_learn==1.2.1
scipy==1.10.0
torch==1.13.1
torchvision==0.14.1
scikit-learn~=1.2.1
opencv-python~=4.7.0.72
scikit-image~=0.21.0